New grep module for rules
This commit adds a new grep module that can be used to search for
patterns in files or directories, recursively or not.

Users will be able to make the module fail if the pattern is found or
is not found.

This module shell-out to the grep CLI because it's more convenient
than reimplementing grep in pure golang.

Alpine and GolangCI lint versions are also upgraded.

Signed-off-by: Mathieu Corbin <mathieu.corbin@qonto.com>