Add support for extended regexp in grep rule
The `grep` rule is using the default behaviour of `grep` to match the
pattern. This default is using [the "basic regular expression"][1],
which can be less powerful in some cases. For instance, `abc?` won’t
match `ab`.

To solve this, `grep` command accepts a [`--extended-regexp` flag][2],
which this patch is adding to the `grep` rule.

[1]: https://man7.org/linux/man-pages/man1/grep.1.html#REGULAR_EXPRESSIONS
[2]: https://man7.org/linux/man-pages/man1/grep.1.html#OPTIONS

Signed-off-by: Maxime Demolin <maxime.demolin@qonto.com>